pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "doom-mcp"
version = "0.1.0"
description = "MCP server that lets AI agents play Doom via ViZDoom"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "vizdoom>=1.2.0",
    "Pillow",
    "numpy",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[tool.hatch.build.targets.wheel]
packages = ["src/doom_mcp"]

[tool.pytest.ini_options]
markers = [
    "integration: tests requiring ViZDoom runtime",
]
asyncio_mode = "auto"
